COSBETA-23: Commonising antivirus between mspsds and cosmetic. Set up the storage initializer for cosmetics.
